weapon 데미지 float으로 수정

+++ Assets/Scripts/Interfaces/IDamagable.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IDamagable
 {
     int Health { get; }
-    void GetDamaged(int damage);
+    void GetDamaged(float damage);
 }
